requirements.txt
streamlit
google-generativeai
python-docx
PyPDF2
langchain-text-splitters
